+++ JDI.Light/JDI.Light.Tests/UIObjects/Pages/Html5Page.cs
@@ -2,6 +2,7 @@
 using JDI.Light.Elements.Common;
 using JDI.Light.Elements.Composite;
 using JDI.Light.Interfaces.Common;
+using JDI.Light.Interfaces.Complex;
 
 namespace JDI.Light.Tests.UIObjects
 {
