+++ JDI.Light/JDI.Light/JDI.Light.csproj
@@ -109,7 +109,7 @@
     <Compile Include="Interfaces\Base\IVisible.cs" />
     <Compile Include="Interfaces\Common\IButton.cs" />
     <Compile Include="Interfaces\Common\ICheckBox.cs" />
-    <Compile Include="Interfaces\Common\ICheckList.cs" />
+    <Compile Include="Interfaces\Complex\ICheckList.cs" />
     <Compile Include="Interfaces\Common\IDataList.cs" />
     <Compile Include="Interfaces\Common\IDatePicker.cs" />
     <Compile Include="Interfaces\Common\IDateTimeSelector.cs" />
